requirements.txt
asgiref==3.5.2
boto3==1.24.67
botocore==1.27.67
certifi==2022.6.15
charset-normalizer==2.1.0
dj-stripe==2.6.2
Django==4.0.5
django-anymail==8.6
django-filter==22.1
django-mailjet==0.3.1
django-robots==5.0
django-storages==1.13.1
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.12.0
jmespath==1.0.1
jsonfield==3.1.0
mailjet-rest==1.3.4
Markdown==3.4.1
Pillow==9.2.0
psycopg2==2.9.3
python-dateutil==2.8.2
python-dotenv==0.21.0
requests==2.28.1
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.2
stripe==4.1.0
urllib3==1.26.10
whitenoise==6.2.0
zipp==3.8.1
